requirements.txt
Flask
openai
requests
beautifulsoup4
python-dotenv
python-wordpress-xmlrpc
